pyproject.toml
[project]
name = "youtube-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb==1.5.7",
    "colorama==0.4.6",
    "faiss-cpu==1.13.2",
    "huggingface-hub==1.9.2",
    "langchain==1.2.15",
    "langchain-anthropic==1.4.0",
    "langchain-classic==1.0.3",
    "langchain-community==0.4.1",
    "langchain-core==1.2.27",
    "langchain-experimental==0.4.1",
    "langchain-google-genai==4.2.1",
    "langchain-groq==1.1.2",
    "langchain-huggingface==1.2.1",
    "langchain-openai==1.1.12",
    "langchain-text-splitters==1.1.1",
    "pandas>=2.3.3",
    "pydantic==2.12.5",
    "python-dotenv==1.2.2",
    "requests==2.33.1",
    "sentence-transformers==5.3.0",
    "streamlit==1.56.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "torch==2.11.0",
    "transformers==5.5.0",
    "typing==3.7.4.3",
    "youtube-transcript-api==1.2.4",
]
